Simplify the use of custom .() and .[]
This commit modifies the parser to use the newly defined .() and .[]
operators. It also moves the definition of the standard .() and .[]
operator for String/Bytes and Array to the pervasives module.

Before this commit, expressions of the form array.(index) and
string.[index] where desugared to Array.get[_unsafe] array index and
String.get[_unsafe] string index. The unsafe or unsafe version were
chosen depending on the presence of the "-unsafe" compiler option.
Such expression are now desugared to ( .() ) array index and ( .[] )
string index respectively. The same desugar operation is applied to
array.(index) <- value which becomes ( .()<- ) array index value.

In order to keep the standard semantic for the string and array index
operations, these new index operators are defined in the pervasives
module using new compiler primitives, e.g.
let .() = "%array_opt_get".
These new primitives are then mapped to safe or unsafe version
depending on the the "-unsafe" compiler option. Consequently, these
modifications should have no impact on existing code.

With these modifications, defining custom .() and .[] operators should
be easier, at the cost of losing access to the standard index operator
for either array or string.