disambiguation for extension constructors
This commit exposes all extension constructors when looking up for
a construction with a given type in the environment.

This makes constructor disambiguation work for extension constructors.

Going one step further, when the name of an extension constructors
is misspelled, we cannot rely on the type to find all possible
names. However, since we are in an error path, we have some
path at hand.
Thus, this commit adds an "extra_spellchecking_candidates" function
that is used to search for more spellchecking candidates in the error
case.
For extension constructors, this means going through all constructors
in scope and picking all constructor with the right type.

This commit should not change anything for labels and standard constructors.