shapes: use strong reduction (reduction under functor abstractions)
Any reduction work not done by the compiler will have to be done by
Merlin. We believe that strong-reduction (including under functors) is
preferable to weak-reduction in the compiler, if we can avoid time and
space blowups.

At this point, irmin.cmt needs 5mn to produce a 2.5Mio .cmt file; this
is too slow, but memory usage is fine.